rename OtherHosts -> PreviousHosts
To ensure consistent naming in code and plugin names, rename all code
reference to OtherHosts to PreviousHosts.
