+++ source/extensions/extensions_build_config.bzl
@@ -113,7 +113,7 @@ EXTENSIONS = {
     "envoy.transport_sockets.capture":                  "//source/extensions/transport_sockets/capture:config",
 
     # Retry host predicates
-    "envoy.retry_host_predicates.other_hosts":          "//source/extensions/retry/host/other_hosts:config",
+    "envoy.retry_host_predicates.previous_hosts":          "//source/extensions/retry/host/previous_hosts:config",
 
     # Retry priorities
     "envoy.retry_priorities.previous_priorities":       "//source/extensions/retry/priority/other_priority:config",
