+++ test/extensions/retry/host/previous_hosts/BUILD
@@ -14,9 +14,9 @@ envoy_package()
 envoy_extension_cc_test(
     name = "config_test",
     srcs = ["config_test.cc"],
-    extension_name = "envoy.retry_host_predicates.other_hosts",
+    extension_name = "envoy.retry_host_predicates.previous_hosts",
     deps = [
-        "//source/extensions/retry/host/other_hosts:config",
+        "//source/extensions/retry/host/previous_hosts:config",
         "//test/mocks/upstream:upstream_mocks",
     ],
 )
